:recycle: [ConsultationCentre] Migrate
